requirements.txt
aiohttp
anytree
appdirs
aria2p
attrdict
beautifulsoup4
cfscrape
cloudscrape
feedparser
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
gunicorn
hachoir
js2py
lk21
lxml
pillow
psutil
psycopg2-binary
pybase64
pyrogram
pyshorteners
python-dotenv
python-magic
python-telegram-bot
qbittorrent-api
requests
speedtest-cli
telegraph
tenacity
TgCrypto
torrentool==1.1.0
urllib3
yt-dlp
pyshorteners
youtube_dl
bs4
natsort
Pillow
pytz
yt-dlp
messages
heroku3
html-telegraph-poster
cfscrape
gitpython
heroku3
html-telegraph-poster
